+++ libft/ft_llintlen.c
@@ -6,15 +6,15 @@
 /*   By: vpetit <vpetit@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/30 17:31:46 by vpetit            #+#    #+#             */
-/*   Updated: 2017/08/30 17:32:41 by vpetit           ###   ########.fr       */
+/*   Updated: 2017/09/26 17:43:08 by vpetit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t		ft_llintlen(long long nbr)
+int			ft_llintlen(long long nbr)
 {
-	size_t	len;
+	int		len;
 
 	len = 1;
 	nbr < 0 ? len++ : len;
